refactor: LottoNumber, LottoNumbers를 shop -> lottonumber package로 이동
