+++ src/main/kotlin/lotto/domain/lottonumber/LottoNumber.kt
@@ -1,4 +1,4 @@
-package lotto.domain.shop
+package lotto.domain.lottonumber
 
 @JvmInline
 value class LottoNumber(val value: Int) : Comparable<LottoNumber> {
