+++ src/main/kotlin/lotto/domain/shop/LottoGame.kt
@@ -1,4 +1,6 @@
 package lotto.domain.shop
 
+import lotto.domain.lottonumber.LottoNumbers
+
 @JvmInline
 value class LottoGame(val lottoNumbers: LottoNumbers)
